Fri, Nov 8, 2024, 10:55 AM +01:00

+++ calculator.c
@@ -3,8 +3,8 @@
 
 int main(void)
 {
-    float x = get_int("x: ");
-    float y = get_int("y: ");
+    double x = get_int("x: ");
+    double y = get_int("y: ");
 
     printf("%.20f\n", x / y);
 }
